% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupPLS.predict.R
\name{groupPLS.predict}
\alias{groupPLS.predict}
\title{Predict for new observations}
\usage{
groupPLS.predict(groupPLS.obj, newdata = newdata)
}
\arguments{
\item{groupPLS.obj}{the object from fitting \code{groupPLS}.}

\item{newdata}{predictors of new observations.}
}
\value{
Matrix of predicted values.
}
\description{
Prediction for new observations from the fit of two-step group penalized partial least squares regression.
}
