% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupPLS.R
\name{groupPLS}
\alias{groupPLS}
\title{Two-step group penalized partial least squares regression}
\usage{
groupPLS(formula, scale.X = FALSE, scale.Y = TRUE, k,
  group.index = NA, group.pen = TRUE, lambda.min = 0.01,
  lambda.max = 0.1, ncomp = 20, seed = 1)
}
\arguments{
\item{formula}{model formula, \code{y~x}, where \code{y} are the response variables and x are the predictors.}

\item{scale.X}{scale \code{x} by the sample standard deviation of each variable. TRUE or FALSE}

\item{scale.Y}{scale \code{y} by the sample standard deviation of each variable. TRUE or FALSE}

\item{k}{number of groups. If \code{group.index} is not provided, then perform the hierarchical clustering based on a similarity matrix to define \code{k} clusters.}

\item{group.index}{pre-defined group index for each predictor.}

\item{group.pen}{group penalization on latent components. TRUE or FALSE.}

\item{lambda.min}{lower bound of group penalty paramter.}

\item{lambda.max}{upper bound of group penalty paramter.}

\item{ncomp}{number of latent components in partial least squares.}

\item{seed}{an integer.}
}
\value{
Return \code{groupPLS} object.

\item{ncomp.group}{indices of latent components from PLS in each group.}

\item{latent.group}{combined latent components from PLS in each group.}

\item{coef}{regression coefficients of latent components on responses.}

\item{scale.X}{if scaling of X is requested.}

\item{scale.Y}{if scaling of Y is requested.}

\item{Xmean}{mean of each variable in X.}

\item{Xsd}{standard deviation of each variable in X.}

\item{Ymean}{mean of each variable in Y.}

\item{Ysd}{standard deviation of each variable in Y.}

\item{k}{number of groups.}

\item{group.index}{group index for each predictor.}

\item{group.lambda}{group penalty paramter selected by cross validation.}

\item{call}{function call.}
}
\description{
Performs group penalized partial least squares regression.
}
\details{
More details.
}
\examples{
data<-read.csv("Spectra_eps.csv")
data<-data[complete.cases(data),]
rs<-groupPLS(data[,c(2,4:6)] ~ data[,7:dim(data)[2]], k=3)

}
\author{
Jiali Wang (\email{jiali.wang@data61.csiro.au}); Le Chang (\email{le.chang@anu.edu.au})
}
