% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vipscore.R
\name{vipscore}
\alias{vipscore}
\title{Variable importance in projection}
\usage{
vipscore(groupPLS.obj, y)
}
\arguments{
\item{groupPLS.obj}{the object from fitting \code{groupPLS}.}

\item{y}{data frame contains response variables.}
}
\value{
Variable importance in projection scores for response variables.
}
\description{
Compute variable importance in projection scores (VIP) for response variables.
}
\examples{
vip<-vipscore(groupPLS.obj=rs,y=ytrain)
par(mfrow=c(3,1))
for (i in 1:3){
plot(vip[,i],type="h")
}

}
